pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "devise-greenness"
version = "1.0.0"
description = "Package for Paper DEVISE : Depth-Aware Vegetation Indexing System"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
  { name = "Cantay Caliskan", email = "ccaliska@ur.rochester.edu" },
  { name = "Zhizhuang Chen", email = "zchen141@u.rochester.edu" },
  { name = "Junjie Zhao", email = "jzhao58@u.rochester.edu" },
  { name = "Linglan Yang", email = "lyang49@u.rochester.edu" },
  { name = "Mingzhen Zhang", email = "mzhang96@u.rochester.edu" }
]
keywords = ["segementation", "Computer Vision", "Urban Analytics", "Green Vegetation Index", "Monocular Depth Estimation", "Computational Social Science"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  "numpy>=1.21.0",
  "opencv-python>=4.5.0",
  "Pillow>=8.0.0",
  "torch>=1.9.0",
  "transformers>=4.15.0"
]

[project.urls]
Homepage = "https://github.com/NNyarlathotep/DEVISE"
Issues = "https://github.com/NNyarlathotep/DEVISE/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/devise_pkg"]
